// tools/conformance/src/modules/nestjs-graphql-module-generator/nestjs-graphql-module-generator.types.ts
// 🏷️ Types

import type { Tree } from "@nx/devkit";

/**
 * Auto-generated documentation placeholder.
 */
export interface NestjsGraphqlModuleGeneratorArguments {
  options: NestjsGraphqlModuleGeneratorOptions;
  tree: Tree;
}

/**
 * Auto-generated documentation placeholder.
 */
export interface NestjsGraphqlModuleGeneratorOptions {
  name: string;
  project?: string;
}